third_party/blink/web_tests/platform/win11-arm64/http/tests/xmlhttprequest/redirect-cross-origin-tripmine-expected.txt

Test that a cross-origin redirect does not result in a non-simple request being sent to the target.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Asynchronous XMLHttpRequest 307 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 307 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 303 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 303 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 302 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 302 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 307 DELETE redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 301 POST redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 301 GET redirect:
PASS tripmineStatus is ""
Asynchronous XMLHttpRequest 301 DELETE redirect:
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 307 POST redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 307 GET redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 303 POST redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 303 GET redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 302 POST redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 302 GET redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 307 DELETE redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 301 POST redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 301 GET redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
Synchronous XMLHttpRequest 301 DELETE redirect:
FAIL testSyncReq.send('data') should throw an exception. Was undefined.
PASS tripmineStatus is ""
PASS successfullyParsed is true

TEST COMPLETE
